pyproject.toml
[tool.poetry]
name = "bananagen"
version = "1.0.0"
description = "CLI tool for generating image assets using Gemini 2.5 Flash"
authors = ["Bananagen Team <info@bananagen.org>"]
readme = "README.md"
packages = [{include = "bananagen"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
pillow = "^10.0.0"
aiohttp = "^3.8.0"
fastapi = "^0.104.0"
pydantic = "^2.0.0"
uvicorn = "^0.24.0"
google-generativeai = "^0.8.0"
cryptography = "^43.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core"

[tool.mypy]
ignore_missing_imports = true

[tool.poetry.scripts]
bananagen = "bananagen.cli:main"
